Resolving vpn not adding search domains
Search domains were not being used to configure systemd-resolved. as a result, .zillow.local domains we're not being resolved in the browser or by ping.

Some deep digging and false starts resulted in the discovering the culprit: the CISCO_DEF_DOMAIN environment variable not being set, resulting in domains not being added to the link entry in systemd-resolved.