awesome: fix losing tags after display disconnect
I had a bug where I would lose some of my tags (awesomewm term for a
view of the applications) when a display disconnects.

This fixes that, as well as fixing slightly broken display defaulting
behavior.

During this I ran into puppet's File directive no longer working on the
~/.config directory, because some applications like Chrome and vscode
adding socket files. So I switched to rsync.

Attempting to upstream the sharedtags fix:
https://github.com/Drauthius/awesome-sharedtags/pull/8